Migrate examples to the new splitter scheme

+++ examples/plot_replay_experiment.py
@@ -118,4 +118,5 @@ value_store.close()
 # We clean all the cache folder.
 
 shutil.rmtree(CACHE_PATH)
-os.remove(DATABASE_PATH)
+if os.path.exists(DATABASE_PATH):
+    os.remove(DATABASE_PATH)
